+++ SampleApp/PPHSDKSampleApp/OptionsCell.swift
@@ -1,0 +1,16 @@
+//
+//  OptionsCell.swift
+//  PPHSDKSampleApp
+//
+//  Created by Priyank Shah on 3/24/21.
+//  Copyright © 2021 cowright. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+
+class OptionsCell: UITableViewCell {
+    @IBOutlet weak var lblOp: UILabel!
+}
+
