+++ SampleApp/PPHSDKSampleApp/TagsCell.swift
@@ -1,0 +1,14 @@
+//
+//  TagsCell.swift
+//  PPHSDKSampleApp
+//
+//  Created by Priyank Shah on 3/24/21.
+//  Copyright © 2021 cowright. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class TagsCell: UITableViewCell {
+    @IBOutlet weak var txtTags: UITextField!
+}
