+++ SampleApp/PPHSDKSampleApp/TransactionOptionsCell.swift
@@ -1,0 +1,15 @@
+//
+//  TransactionOptionsCell.swift
+//  PPHSDKSampleApp
+//
+//  Created by Priyank Shah on 3/24/21.
+//  Copyright © 2021 cowright. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class TransactionOptionsCell: UITableViewCell {
+    @IBOutlet weak var lblOption: UILabel!
+    @IBOutlet weak var switchOption: UISwitch!
+}
